+++ GerenciadorAmbiente/src/main/java/jeffersonmca/com/github/gerenciadorambiente/servico/ServicoPessoa.java
@@ -1,7 +1,10 @@
 package jeffersonmca.com.github.gerenciadorambiente.servico;
 
 import java.util.List;
+
+import jeffersonmca.com.github.gerenciadorambiente.modelo.Periodo;
 import jeffersonmca.com.github.gerenciadorambiente.modelo.Pessoa;
+import jeffersonmca.com.github.gerenciadorambiente.validacao.Validacao;
 import jeffersonmca.com.github.gerenciadorambiente.dao.DAOPessoa;
 import jeffersonmca.com.github.gerenciadorambiente.excecoes.ExcecaoServico;
 
@@ -27,15 +30,27 @@ public class ServicoPessoa {
     public Pessoa getPessoa(Integer codigo) throws ExcecaoServico {
         
         /*Regra de negocio*/
-        
-        return dao.getPessoa(codigo);
+    	if (Validacao.Identificador(codigo)) {
+    		return dao.getPessoa(codigo);
+    	}
+    	
+    	return null;
     }
     
     
     public Pessoa remover(Integer codigo) throws ExcecaoServico {
         
         /*Regra de negocio*/
-              
-        return dao.remover(codigo);
+    	if (Validacao.Identificador(codigo)) {
+            
+            Pessoa aux = dao.getPessoa(codigo);
+            
+            // Esta no BD?
+            if (Validacao.Alocado(aux)) {
+                return dao.remover(codigo);
+            }
+        }
+        
+        return null;
     } 
 }
